requirements.txt
request
fastapi
jinja2
uvicorn